Separate version info from system information
