+++ utilities/src/main/java/org/craftercms/commons/monitoring/VersionInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2026 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -19,9 +19,6 @@ import org.apache.commons.lang3.StringUtils;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.management.ManagementFactory;
-import java.lang.management.OperatingSystemMXBean;
-import java.lang.management.RuntimeMXBean;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.time.Instant;
@@ -31,17 +28,17 @@ import java.util.jar.Attributes;
 import java.util.jar.JarFile;
 import java.util.jar.Manifest;
 
+import static org.apache.commons.lang3.Strings.CS;
+
 /**
  * Get's current version information about:
- * 1. Current VM version runs current code.
- * 2. Build Information from a given manifest file.
- * 3. Name and version from a given manifest file.
+ * 1. Build Information from a given manifest file.
+ * 2. Name and version from a given manifest file.
  *
  * @author Carlos Ortiz
  * @since 3.0
  */
-public final class VersionInfo {
-
+public class VersionInfo {
 	private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ISO_INSTANT.withZone(ZoneId.of("UTC"));
 
 	/**
@@ -84,39 +81,10 @@ public final class VersionInfo {
 	private String packageBuild;
 	private String packageBuildDate;
 
-	private String osName;
-	private String osVersion;
-	private String osArch;
 
-	private String javaVersion;
-	private String javaVendor;
-	private String javaVm;
-
-	/**
-	 * Create a VersionInfo pojo instance based on a given Manifest File.
-	 * Empty values are assign if the keys in the manifest are missing.
-	 *
-	 * @param manifest Manifest File that contains the VersionInfo Information.
-	 */
-	private VersionInfo(Manifest manifest) {
+	protected VersionInfo(Manifest manifest) {
 		Attributes mainAttrs = manifest.getMainAttributes();
 		initFromManifest(mainAttrs);
-		initRuntime();
-		initOS();
-	}
-
-	private void initFromManifest(Attributes mainAttrs) {
-
-		if (mainAttrs.getValue(KEY_BUILD_ON) != null) {
-			packageBuildDate = FORMATTER.format(
-				Instant.ofEpochMilli(Long.parseLong(mainAttrs.getValue(KEY_BUILD_ON))));
-		} else {
-			packageBuildDate = StringUtils.EMPTY;
-		}
-		packageName = mainAttrs.getValue(KEY_IMPLEMENTATION_TITLE);
-		packageVersion = mainAttrs.getValue(KEY_IMPLEMENTATION_VERSION);
-		packageBuild = mainAttrs.getValue(KEY_IMPLEMENTATION_BUILD);
-
 	}
 
 	/**
@@ -136,56 +104,46 @@ public final class VersionInfo {
 	 * @return A {@link VersionInfo} pojo with the information.
 	 * @throws IOException If Unable to read the Manifest file.
 	 */
-	public static VersionInfo getVersion(Class clazz) throws IOException {
+	public static VersionInfo getVersion(Class<?> clazz) throws IOException {
+		Manifest manifest = findManifest(clazz);
+		if (manifest == null) {
+			return null;
+		}
+		return new VersionInfo(manifest);
+	}
+
+	protected static Manifest findManifest(Class<?> clazz) throws IOException {
 		String path = clazz.getProtectionDomain().getCodeSource().getLocation().getPath();
 		if (StringUtils.isEmpty(path)) {
 			return null;
 		}
-		if (StringUtils.contains(path, WAR_PATH)) {
-			path = StringUtils.removeEnd(path, WAR_PATH);
-			path = StringUtils.appendIfMissing(path, MANIFEST_PATH);
+		if (CS.contains(path, WAR_PATH)) {
+			path = CS.removeEnd(path, WAR_PATH);
+			path = CS.appendIfMissing(path, MANIFEST_PATH);
 			try (InputStream is = Files.newInputStream(Paths.get(path))) {
-				return getVersion(new Manifest(is));
+				return new Manifest(is);
 			}
-		} else if (StringUtils.contains(path, SPRING_PATH)) {
-			path = StringUtils.removeStart(path, "file:");		// filesystem case
-			path = StringUtils.removeStart(path, "nested:"); 	// Spring Boot nested jars case
-			path = StringUtils.removeEnd(path, SPRING_PATH);
+		}
+		if (CS.contains(path, SPRING_PATH)) {
+			path = CS.removeStart(path, "file:");        // filesystem case
+			path = CS.removeStart(path, "nested:");    // Spring Boot nested jars case
+			path = CS.removeEnd(path, SPRING_PATH);
 			JarFile jarFile = new JarFile(path);
-			return VersionInfo.getVersion(jarFile.getManifest());
+			return jarFile.getManifest();
 		}
 		return null;
 	}
 
-	private void initOS() {
-		OperatingSystemMXBean os = ManagementFactory.getOperatingSystemMXBean();
-		osName = os.getName();
-		osVersion = os.getVersion();
-		osArch = os.getArch();
-	}
-
-	@SuppressWarnings("unchecked")
-	private void initRuntime() {
-		RuntimeMXBean runtime = ManagementFactory.getRuntimeMXBean();
-		javaVersion = runtime.getSpecVersion();
-		javaVendor = runtime.getSpecVendor();
-		javaVm = runtime.getVmName();
-	}
-
-	@Override
-	public String toString() {
-		return "VersionInfo{" +
-			", packageName='" + packageName + '\'' +
-			", packageVersion='" + packageVersion + '\'' +
-			", packageBuild='" + packageBuild + '\'' +
-			", packageBuildDate='" + packageBuildDate + '\'' +
-			", javaVersion='" + javaVersion + '\'' +
-			", javaVendor='" + javaVendor + '\'' +
-			", javaVm='" + javaVm + '\'' +
-			", osName='" + osName + '\'' +
-			", osVersion='" + osVersion + '\'' +
-			", osArch='" + osArch + '\'' +
-			'}';
+	private void initFromManifest(Attributes mainAttrs) {
+		if (mainAttrs.getValue(KEY_BUILD_ON) != null) {
+			packageBuildDate = FORMATTER.format(
+					Instant.ofEpochMilli(Long.parseLong(mainAttrs.getValue(KEY_BUILD_ON))));
+		} else {
+			packageBuildDate = StringUtils.EMPTY;
+		}
+		packageName = mainAttrs.getValue(KEY_IMPLEMENTATION_TITLE);
+		packageVersion = mainAttrs.getValue(KEY_IMPLEMENTATION_VERSION);
+		packageBuild = mainAttrs.getValue(KEY_IMPLEMENTATION_BUILD);
 	}
 
 	public String getPackageName() {
@@ -203,30 +161,4 @@ public final class VersionInfo {
 	public String getPackageBuildDate() {
 		return packageBuildDate;
 	}
-
-	public String getJavaVersion() {
-		return javaVersion;
-	}
-
-	public String getJavaVendor() {
-		return javaVendor;
-	}
-
-	public String getJavaVm() {
-		return javaVm;
-	}
-
-	public String getOsName() {
-		return osName;
-	}
-
-	public String getOsVersion() {
-		return osVersion;
-	}
-
-	public String getOsArch() {
-		return osArch;
-	}
-
 }
-
