+++ utilities/src/main/java/org/craftercms/commons/monitoring/rest/MonitoringRestControllerBase.java
@@ -16,10 +16,10 @@
 
 package org.craftercms.commons.monitoring.rest;
 
-import org.apache.commons.lang3.StringUtils;
 import org.craftercms.commons.exceptions.InvalidManagementTokenException;
 import org.craftercms.commons.monitoring.MemoryInfo;
 import org.craftercms.commons.monitoring.StatusInfo;
+import org.craftercms.commons.monitoring.SysInfo;
 import org.craftercms.commons.monitoring.VersionInfo;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,6 +27,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.io.IOException;
+
+import static org.apache.commons.lang3.Strings.CS;
 
 /**
  * Base controller for all monitoring related APIs
@@ -40,6 +42,7 @@ public abstract class MonitoringRestControllerBase {
 	public final static String MEMORY_URL = "/memory";
 	public final static String STATUS_URL = "/status";
 	public final static String VERSION_URL = "/version";
+	public final static String SYSINFO_URL = "/sysinfo";
 	public final static String DISK_URL = "/disk";
 
 	private final String configuredToken;
@@ -49,29 +52,36 @@ public abstract class MonitoringRestControllerBase {
 	}
 
 	@GetMapping(ROOT_URL + MEMORY_URL)
-	public MemoryInfo getCurrentMemory(@RequestParam(name = "token", required = true) String token)
-		throws InvalidManagementTokenException {
+	public MemoryInfo getCurrentMemory(@RequestParam(name = "token") String token)
+			throws InvalidManagementTokenException {
 		validateToken(token);
 		return MemoryInfo.getCurrentMemory();
 	}
 
 	@GetMapping(ROOT_URL + STATUS_URL)
 	public ResponseEntity getCurrentStatus(@RequestParam(name = "site", required = false) String site,
-					       @RequestParam(name = "token") String token)
-		throws InvalidManagementTokenException {
+										   @RequestParam(name = "token") String token)
+			throws InvalidManagementTokenException {
 		validateToken(token);
 		return ResponseEntity.ok().body(StatusInfo.getCurrentStatus());
 	}
 
 	@GetMapping(ROOT_URL + VERSION_URL)
-	public ResponseEntity getCurrentVersion(@RequestParam(name = "token", required = true) String token)
-		throws InvalidManagementTokenException, IOException {
+	public ResponseEntity getCurrentVersion(@RequestParam(name = "token") String token)
+			throws InvalidManagementTokenException, IOException {
 		validateToken(token);
 		return ResponseEntity.ok().body(VersionInfo.getVersion(this.getClass()));
 	}
 
+	@GetMapping(ROOT_URL + SYSINFO_URL)
+	public ResponseEntity getCurrentSysInfo(@RequestParam(name = "token") String token)
+			throws InvalidManagementTokenException, IOException {
+		validateToken(token);
+		return ResponseEntity.ok().body(SysInfo.getInfo(this.getClass()));
+	}
+
 	protected final void validateToken(final String requestToken) throws InvalidManagementTokenException {
-		if (!StringUtils.equals(requestToken, configuredToken)) {
+		if (!CS.equals(requestToken, configuredToken)) {
 			throw new InvalidManagementTokenException("Management authorization failed, invalid token.");
 		}
 	}
